- рефакторинг сортировки
